requirements.txt
yara-python~=4.3.1
vt-py~=0.17.5
pefile~=2023.2.7